Only update state if component is still mounted
`setIsLoading()` comes from `useState()` which will trigger a rerender when called. If
the component has been unmounted already, this will throw a warning
about memory leaks. We only really want to call `setIsLoading()` if the
component is still mounted. So we'll `useRef` to a boolean to keep track
of weather the component `isMounted`, and use a `useEffect` to update
`isMounted` when the component is unmounted.